PNAS 01/24/20_allow pb happen in the center location
The current problem is:
In the MDP setup, when pb, no change in beliefs, reward would not go from 1 to zero. After pb, wait for one time-step.
The problem is that, ex. loc=2, rew[t-1] = 1, pb = 4, rew[t] = 0. Using p(b'|b, o, a) would cause a problem since this is something that would not happen.
Before, in the Baum-Welch, for the button pressing side (and the other side), the hybrid transition uses ThA directly, this allows probability of rew going from 1 to 0.
2 solutions:
(1) after pb,the action is NA. Pb does not change beliefs. One thing needs to pay attention is that in the Baum-Welch, when NA is after pb, the policy is deterministic.
(2) change p(b'|b, o, a) by a allowing one step of doing nothing. This may cause discrepancy? Need to marginalize over color for the intermediate time point.